Backend/requirements.txt
aioredis==1.3.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.7
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-google-maps==0.13.0
django-heroku==0.3.1
django-mysql==4.7.0
django-rest-framework==0.1.0
django-tinymce==3.4.0
djangorestframework==3.13.1
engineering-notation==0.6.0
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.7.1
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
install==1.3.5
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.9
jsonschema==4.9.1
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
MouseInfo==0.1.3
msgpack==1.0.4
mysqlclient==2.1.1
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.2.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.9.1
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.12.0
PyJWT==2.4.0
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.18.1
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytweening==1.0.4
pytz==2022.1
pywin32==304
pywinpty==2.0.7
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.4.0
Send2Trash==1.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
terminado==0.15.0
tinycss2==1.1.1
tk==0.1.0
tk-tools==0.15.0
toml==0.10.2
tornado==6.2
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
whitenoise==6.1.0
wsproto==1.1.0
zope.interface==5.4.0
